refactor: move autoFarm functions into autoFarmHelpers
